first wave of renaming
